Added instantiateLoner.
instantiateLoner is a way to leak Instance out of Hierarchies for
entities that don’t depend on others. That is, for loner entities.
That function is important because a LOT of entities will be loner
ones, so not using Hierarchy is a plus here for performance
purposes.

Partial fix for #99.