Glyph type added.
A Glyph represents a character replaced by meta-data. Those meta-data
are used to perform correct lookups in the SDF fontmap.

That commit also introduces the concept of « line buffer ». The
text renderer will be granted the handle of an OpenGL buffer that
will represent a temporary storage for a line of a text to render.
We might need to be able to make Glyph an instance of a typeclass we
don’t have yet: something telling us we can send it into a GLSL
block (std430).

Related: #106.